Remove unused JS libs, reimplement jTruncate()
