+++ root/footer.mhtml
@@ -7,7 +7,6 @@
 % if ($c->debug()) {
 <script src="<% $c->uri_for('/static/js/jquery-v1.9.1.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/jquery-ui-v1.9.2.js') %>"></script>
-<script src="<% $c->uri_for('/static/js/jquery.validate-v1.17.0.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/angular-v1.7.8.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/bootstrap-3.3.7.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/angular-animate-v1.5.11.js') %>"></script>
@@ -18,7 +17,6 @@
 % } else {
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
 <script src="//ajax.googleapis.com/ajax/libs/jqueryui/1.9.2/jquery-ui.min.js"></script>
-<script src="//ajax.microsoft.com/ajax/jquery.validate/1.17.0/jquery.validate.min.js"></script>
 <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.7.8/angular.min.js"></script>
 <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
 <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.5.11/angular-animate.min.js"></script>
@@ -27,10 +25,7 @@
 <script src="//cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.2/Chart.min.js"></script>
 <script src="//cdnjs.cloudflare.com/ajax/libs/angular-chart.js/1.1.1/angular-chart.min.js"></script>
 % }
-<script src="<% $c->uri_for('/static/js/jquery.shiftcheckbox.0.8.0.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/jquery.pnotify.min.js') %>"></script>
-<script src="<% $c->uri_for('/static/js/jquery.placeholder.min.js') %>"></script>
-<script src="<% $c->uri_for('/static/js/jquery.jtruncate.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/tag-it.js') %>"></script>
 <script src="<% $c->uri_for('/static/js/angular-confirm-1.2.6.min.js') %>"></script>
 <script type="text/javascript" src="<% $c->uri_for('/static/js/canto.js') %>"></script>
@@ -52,12 +47,6 @@
 
   </script>
 % }
-
-<script language="javascript" type="text/javascript">
-$(function() {
-  $('input, textarea').placeholder();
-});
-</script>
 
 <%init>
 my $site_tracking_id = $c->config()->{google_analytics_id};
